Testing the retrieval of the customer and species object of a project
I am testing that the customer and species fields use DBRef correctly. At the current state, the customer field is populated with the customer object corresponding to the objectID when retrieving a Project document, but the species field is NOT populated. This issue may be due to the "Specie"/"Species" namespace conflict?? I have filed a bug to track this issue.