Perform step on plates (controller logic)
Incrementing (moving plates to the next step) now uses the
async module doing away with waiting for an arbitrary amount
of time when all plates are not finished incrementing/unassigning.

The problem with waiting an arbitrary amount of time is that
when only a few plates were to be "incremented", there was a delay between
closing the dialog and redirecting back to the worklist.

The async module guarantees to redirect back to the worklist
as soon as all plates are finished being incremented/unassigned.

Note: I think a progress circle should be displayed when we are
in the process of incrementing/unassigning plates, because if for
some reason doing so is taking a long time, the page will not be
redirected and the user will be looking at a useless page until all
plates are finished incrementing/unassigning.